+++ crates/warpui_core/src/runtime/mod.rs
@@ -187,186 +187,6 @@ where
     }
 }
 
-/// Owns raw mode and alternate-screen state independently of a terminal writer.
-pub struct TuiTerminalGuard(RawModeGuard<CrosstermModeControl>);
-
-impl TuiTerminalGuard {
-    /// Enters raw mode and the alternate screen.
-    pub fn enter() -> io::Result<Self> {
-        Ok(Self(RawModeGuard::enter(CrosstermModeControl)?))
-    }
-}
-
-/// Keeps an app-owned TUI driver and terminal-mode guard alive.
-pub struct TuiDriverHandle {
-    _task: ForegroundTask,
-    _reader: thread::JoinHandle<()>,
-    _guard: TuiTerminalGuard,
-}
-
-struct AppTuiScreen<T: TuiView> {
-    window_id: WindowId,
-    root_view: ViewHandle<T>,
-    presenter: TuiPresenter,
-    renderer: TuiFrameRenderer,
-    terminal: CrosstermTerminal,
-}
-
-impl<T: TuiView> AppTuiScreen<T> {
-    fn new(window_id: WindowId, root_view: ViewHandle<T>) -> Self {
-        Self {
-            window_id,
-            root_view,
-            presenter: TuiPresenter::new(),
-            renderer: TuiFrameRenderer::new(),
-            terminal: CrosstermTerminal::new(),
-        }
-    }
-
-    fn draw(&mut self, ctx: &mut AppContext) -> io::Result<()> {
-        let size = self.terminal.size()?;
-        let area = TuiRect::new(0, 0, size.width, size.height);
-        let invalidation = ctx.take_all_invalidations_for_window(self.window_id);
-        self.presenter
-            .invalidate(&invalidation, ctx, self.window_id);
-        let frame = self.presenter.present(ctx, &self.root_view, area);
-        let mut writer = self.terminal.writer();
-        self.renderer.draw(&mut writer, &frame.buffer, frame.cursor)
-    }
-
-    fn dispatch_event(&mut self, ctx: &mut AppContext, event: &Event) -> bool {
-        if let Event::KeyDown {
-            keystroke,
-            is_composing,
-            ..
-        } = event
-        {
-            let responder_chain = ctx.get_responder_chain(self.window_id);
-            match ctx.dispatch_keystroke(self.window_id, &responder_chain, keystroke, *is_composing)
-            {
-                Ok(true) => return true,
-                Ok(false) => {}
-                Err(error) => log::error!("error dispatching TUI keystroke: {error}"),
-            }
-        }
-
-        let Some(element) = self.presenter.last_element.as_mut() else {
-            return false;
-        };
-        let size = self.terminal.size().unwrap_or_default();
-        let area = TuiRect::new(0, 0, size.width, size.height);
-        let root_view_id = self.root_view.id();
-        let mut event_ctx = TuiEventContext::default();
-        event_ctx.set_origin_view(Some(root_view_id));
-        let handled = {
-            let app = &*ctx;
-            let mut layout_ctx = TuiLayoutContext {
-                rendered_views: &mut self.presenter.rendered_views,
-            };
-            element.dispatch_event(event, area, &mut event_ctx, &mut layout_ctx, app)
-        };
-        for view_id in event_ctx.take_notified() {
-            ctx.notify_view_observers(self.window_id, view_id);
-        }
-        for action in event_ctx.take_typed_actions() {
-            ctx.dispatch_typed_action_for_view(
-                self.window_id,
-                action.origin_view_id,
-                action.action.as_ref(),
-            );
-        }
-        handled
-    }
-}
-
-/// Starts an invalidation-driven TUI driver inside Warp's shared app loop.
-pub fn spawn_tui_driver<T: TuiView>(
-    ctx: &mut AppContext,
-    window_id: WindowId,
-    root_view: ViewHandle<T>,
-) -> io::Result<TuiDriverHandle> {
-    let guard = TuiTerminalGuard::enter()?;
-    let screen = Rc::new(RefCell::new(AppTuiScreen::new(window_id, root_view)));
-    {
-        let screen = screen.clone();
-        ctx.on_window_invalidated(window_id, move |_, ctx| {
-            let Ok(mut screen) = screen.try_borrow_mut() else {
-                return;
-            };
-            if let Err(error) = screen.draw(ctx) {
-                log::error!("failed to draw a TUI frame: {error}");
-            }
-        });
-    }
-    if let Err(error) = screen.borrow_mut().draw(ctx) {
-        log::error!("failed to draw the initial TUI frame: {error}");
-    }
-
-    let weak_app = ctx.weak_app();
-    let (sender, receiver) = async_channel::unbounded::<CrosstermEvent>();
-    let screen_for_input = screen.clone();
-    let reader = thread::Builder::new()
-        .name("warp-tui-input".to_owned())
-        .spawn(move || loop {
-            match event::read() {
-                Ok(event) => {
-                    if block_on(sender.send(event)).is_err() {
-                        break;
-                    }
-                }
-                Err(error) => {
-                    log::error!("failed to read a terminal event: {error}");
-                    break;
-                }
-            }
-        })?;
-
-    let task = ctx.foreground_executor().spawn(async move {
-        while let Ok(event) = receiver.recv().await {
-            let Some(app) = weak_app.upgrade() else {
-                break;
-            };
-            let mut ctx = app.borrow_mut();
-            if is_ctrl_c(&event) {
-                ctx.terminate_app(TerminationMode::ForceTerminate, None);
-                continue;
-            }
-            let Some(event) = crossterm_event_to_warp_event(event) else {
-                continue;
-            };
-            let handled = screen_for_input
-                .borrow_mut()
-                .dispatch_event(&mut ctx, &event);
-            if handled {
-                // Redraws are invalidation-driven: `update_windows` fires the
-                // window's invalidation callback (registered above), which draws.
-                // It normally runs at the end of a flush, but input dispatch here
-                // doesn't flush app effects, and a pure-element scroll never marks
-                // a view dirty — so without this the redraw would be deferred until
-                // some later, unrelated flush (e.g. PTY output), making scrolling
-                // feel laggy and intermittent. Invalidate then drive `update_windows`
-                // directly, mirroring the repaint-task pattern in `app.rs`.
-                ctx.invalidate_all_views();
-                ctx.update_windows();
-            }
-        }
-    });
-
-    Ok(TuiDriverHandle {
-        _task: task,
-        _reader: reader,
-        _guard: guard,
-    })
-}
-
-fn is_ctrl_c(event: &CrosstermEvent) -> bool {
-    matches!(
-        event,
-        CrosstermEvent::Key(key)
-            if key.code == KeyCode::Char('c') && key.modifiers.contains(KeyModifiers::CONTROL)
-    )
-}
-
 impl<T, R> TuiRuntime<T, R>
 where
     T: TuiView,
